requirements.txt
absl-py==0.2.0
astor==0.6.2
bleach==1.5.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
definitions==0.2.0
Flask==1.0.2
gast==0.2.0
grpcio==1.11.0
html5lib==0.9999999
idna==2.6
itsdangerous==0.24
Jinja2==2.10
lxml==4.2.1
Markdown==2.6.11
MarkupSafe==1.0
neo4j-driver==1.6.1
neotime==1.0.0
nltk==3.2.5
numpy==1.14.3
protobuf==3.5.2.post1
pytz==2018.5
PyYAML==3.12
requests==2.18.4
sets==0.3.2
six==1.11.0
tensorboard==1.8.0
tensorflow==1.8.0
termcolor==1.1.0
urllib3==1.22
Werkzeug==0.14.1
